Dashboard_Backend/requirements.txt
flask
flask-cors 
pyodbc 
gunicorn 
python-dotenv
